Big Data Lab/Assignment 4/requirements.txt
googleapis-common-protos
google-cloud-storage
google-cloud-pubsub
google-cloud